+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 * 0.3.3 (xx-xx-2013)
 
   * Add convenient Stream::getDimensions method to extract video dimension.
+  * Add DisplayRatioFixer Frame filter.
 
 * 0.3.2 (08-08-2013)